move SpringRow, spring_column_* to pipeline
