+++ springql-core/src/pipeline.rs
@@ -20,7 +20,8 @@ use anyhow::anyhow;
 use std::{collections::HashSet, sync::Arc};
 
 use crate::error::{Result, SpringError};
-use crate::stream_engine::SinkRow;
+use crate::stream_engine::{SinkRow, SqlValue};
+use crate::SpringValue;
 
 use self::{
     name::StreamName, pipeline_graph::PipelineGraph, pipeline_version::PipelineVersion,
@@ -29,14 +30,82 @@ use self::{
 };
 
 pub use crate::api::low_level_rs::{
-    spring_column_bool, spring_column_f32, spring_column_i16, spring_column_i32, spring_column_i64,
-    spring_column_text, spring_command, spring_open, spring_pop, spring_pop_non_blocking,
-    SpringPipeline,
+    spring_command, spring_open, spring_pop, spring_pop_non_blocking, SpringPipeline,
 };
+
+/// Get an integer column.
+///
+/// # Failure
+///
+/// - [SpringError::Unavailable](crate::error::SpringError::Unavailable) when:
+///   - `i_col` already fetched.
+///   - `i_col` out of range.
+/// - [SpringError::Null](crate::error::SpringError::Null) when:
+///   - Column value is NULL
+pub fn spring_column_i32(row: &SpringRow, i_col: usize) -> Result<i32> {
+    spring_column_not_null(row, i_col)
+}
+
+/// Get a 2-byte integer column.
+///
+/// # Failure
+///
+/// Same as [spring_column_i32()](spring_column_i32)
+pub fn spring_column_i16(row: &SpringRow, i_col: usize) -> Result<i16> {
+    spring_column_not_null(row, i_col)
+}
+
+/// Get a 8-byte integer column.
+///
+/// # Failure
+///
+/// Same as [spring_column_i32()](spring_column_i32)
+pub fn spring_column_i64(row: &SpringRow, i_col: usize) -> Result<i64> {
+    spring_column_not_null(row, i_col)
+}
+
+/// Get a float column.
+///
+/// # Failure
+///
+/// Same as [spring_column_i32()](spring_column_i32)
+pub fn spring_column_f32(row: &SpringRow, i_col: usize) -> Result<f32> {
+    spring_column_not_null(row, i_col)
+}
+
+/// Get a boolean column.
+///
+/// # Failure
+///
+/// Same as [spring_column_i32()](spring_column_i32)
+pub fn spring_column_bool(row: &SpringRow, i_col: usize) -> Result<bool> {
+    spring_column_not_null(row, i_col)
+}
+
+/// Get a text column.
+///
+/// # Failure
+///
+/// Same as [spring_column_i32()](spring_column_i32)
+pub fn spring_column_text(row: &SpringRow, i_col: usize) -> Result<String> {
+    spring_column_not_null(row, i_col)
+}
+
+fn spring_column_not_null<T: SpringValue>(row: &SpringRow, i_col: usize) -> Result<T> {
+    let v = row.0.get_by_index(i_col)?;
+    if let SqlValue::NotNull(v) = v {
+        v.unpack()
+    } else {
+        Err(SpringError::Null {
+            stream_name: row.0.stream_name().clone(),
+            i_col,
+        })
+    }
+}
 
 /// Row object from an in memory queue.
 #[derive(Debug)]
-pub struct SpringRow(SinkRow);
+pub struct SpringRow(pub(crate) SinkRow);
 
 impl From<SinkRow> for SpringRow {
     fn from(sink_row: SinkRow) -> Self {
